rename authentication file into auth
